Split mathics.core.assignment into  mathics.core.assignment and mathics.eval.assignment (#1264)
And adding docstrings and type annotations.

---------
