+++ mathics/eval/assignments/__init__.py
@@ -1,0 +1,11 @@
+"""
+Evaluation routines and associated code for Built-in function found under module
+mathics.builtins.assignments.
+"""
+
+from mathics.eval.assignments.assignment import ASSIGNMENT_FUNCTION_MAP, eval_assign
+
+__all__ = [
+    "ASSIGNMENT_FUNCTION_MAP",
+    "eval_assign",
+]
